+++ src/dimension/mod.rs
@@ -6,6 +6,7 @@ mod model;
 pub use algebra::{DivideDimension, MultiplyDimension};
 pub use model::{
     AbsoluteTemperatureSemantics, AngleSemantics, BaseSemantics, Dimension,
-    MechanicalImpedanceSemantics, MolarConcentrationSemantics, ReciprocalVolumeSemantics,
-    SpringStiffnessSemantics, SurfaceTensionSemantics, TemperatureDifferenceSemantics,
+    FlexuralRigiditySemantics, MechanicalImpedanceSemantics, MolarConcentrationSemantics,
+    ReciprocalVolumeSemantics, SpringStiffnessSemantics, SurfaceTensionSemantics,
+    TemperatureDifferenceSemantics,
 };