+++ src/systems/si/dimensions.rs
@@ -3,9 +3,9 @@
 use typenum::{N1, N2, N3, N4, N7, P1, P2, P3, P4, Z0};
 
 use crate::dimension::{
-    AbsoluteTemperatureSemantics, AngleSemantics, Dimension, MechanicalImpedanceSemantics,
-    MolarConcentrationSemantics, ReciprocalVolumeSemantics, SpringStiffnessSemantics,
-    SurfaceTensionSemantics, TemperatureDifferenceSemantics,
+    AbsoluteTemperatureSemantics, AngleSemantics, Dimension, FlexuralRigiditySemantics,
+    MechanicalImpedanceSemantics, MolarConcentrationSemantics, ReciprocalVolumeSemantics,
+    SpringStiffnessSemantics, SurfaceTensionSemantics, TemperatureDifferenceSemantics,
 };
 
 /// Dimensionless quantity.
@@ -90,12 +90,16 @@ pub type ElectricPotentialPerPressure = Dimension<P3, Z0, N1, N1, Z0, Z0, Z0>;
 pub type LengthPerElectricPotential = Dimension<N1, N1, P3, P1, Z0, Z0, Z0>;
 /// Surface electric-charge density.
 pub type SurfaceChargeDensity = Dimension<N2, Z0, P1, P1, Z0, Z0, Z0>;
+/// Volume electric-charge density.
+pub type VolumeChargeDensity = Dimension<N3, Z0, P1, P1, Z0, Z0, Z0>;
 /// Mechanical spring stiffness, force per length.
 pub type SpringStiffness = Dimension<Z0, P1, N2, Z0, Z0, Z0, Z0, SpringStiffnessSemantics>;
 /// Mechanical damping coefficient, force times time per length.
 pub type DampingCoefficient = Dimension<Z0, P1, N1, Z0, Z0, Z0, Z0>;
 /// Mechanical impedance, force per velocity, with coherent unit `kg/s`.
 pub type MechanicalImpedance = Dimension<Z0, P1, N1, Z0, Z0, Z0, Z0, MechanicalImpedanceSemantics>;
+/// Flexural rigidity, with coherent unit `J`.
+pub type FlexuralRigidity = Dimension<P2, P1, N2, Z0, Z0, Z0, Z0, FlexuralRigiditySemantics>;
 /// Pressure gradient, pressure per length.
 pub type PressureGradient = Dimension<N2, P1, N2, Z0, Z0, Z0, Z0>;
 /// Hydraulic resistance, pressure per volumetric flow rate.